+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.0.3 (April 08, 2016)
+I've been working hard to make Composer install more efficient for developers :)
+Configuration files are auto-generated via composer install command.
+
 ## v0.0.2 (March 29, 2016)
 _It's been a while..._
 Capistrano integration is now available. Not yet full optimised, but works efficiently.